market_maker_bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.0
pydantic==2.5.0
requests==2.31.0
web3>=6.0.0
eth-account>=0.8.0
python-dotenv==1.0.1
